backend/requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-docx
